Add compare functionality for weekly automated cycling experiments (#10)
The changes in this branch have been used in the weekly automated cycling for several weeks.  The main addition is to compare nc files produced in the current experiment (background model states, background model state statistics, omf statistics, omb+oma statistics) to those from a benchmark experiment.

The benchmark experiment directory is controlled in `config/benchmark.csh`.  That functionality can be used outside of the automated testing by turning on either of `CompareDA2Benchmark` or `CompareBG2Benchmark` in `drive.csh`.  These are turned off by default.

For those interested, modification of the MPAS-Workflow configuration for the automated testing is controlled in a separate repository (https://github.com/svahl991/mpas-jedi-autotest).