Delete zeroth forecast and diag files (#19)
The zero duration forecast and diagnostics files from the MPAS-Atmosphere integration are not presently used in cycling.  They account for roughly half the disk space in the CyclingFC directory, which is substantial as the number of ensemble members grows.  This PR removes those unneeded files.

The zero duration forecast files from extended forecasts are used for 0-hr (analysis state) verification, and so cannot be deleted.  However, they account for relatively less disk space.