Create new variationalPrep.csh for VariationalIterations and Ensemble Pb, improve StaticFieldsFile handling (#31)
7183787 and d7688c7
Some effort is made to reduce the EDA wall-time caused by concurrent access (across EDA members) of a single static.nc file via the static stream.  Now each EDA member gets its own static file.  There is not much impact to the wall-time, but this is no longer an unknown in terms of performance.


c93512e
The YAML sed substitution for VariationalIterations and all Ensemble B components is moved from jediPrep.csh (used for both hofx and variational) to variationalPrep.csh.

Addresses the second bullet in #20.

The sed substitution for `AppType` (i.e., `hofx` or `variational`) in the observation feedback file names is removed from the `ObsPlugs` and instead included in the file prefix substitution within `jediPrep.csh`.  The goal is to improve transparency in file naming conventions.