Independent EDA member jobs (#32)
### Commit descriptions
096757f
Changes from using relative directories to absolute directories in namelist.atmopshere, streams.atmosphere, and most YAML files.  The purpose is to allow for applications (only Variational for now) to be run in a different directory than the run directory.  This facilitates extending the Variational application to an ensemble of running tasks more simple.

57109f0
Several scripts are renamed for clarity, and the changes are cascaded through the rest of the workflow
 - `clean-hofx.csh` -> `CleanHofX.csh`
 - `clean-variational.csh` -> `CleanVariational.csh`
 - `hofx.csh` -> `HofX.csh`
 - `jediPrep.csh` -> `PrepJEDI.csh`
 - `variationalPrep.csh` -> `PrepVariational.csh`
 - `variational.csh` -> `Variational.csh`

cc33314
Addresses third bullet of #20.
Move background state preparation from `Variational` (formerly `variational`) to `PrepVariational` (formerly `variationalPrep`) to reduce wall-time in CyclingDA task.

e2b0799
Very small clarification to a comment

2ca1535
Small refactoring in HofX and RTPP that will also be used in the Variational scripts in the next commit

b43fc56 + ddc2a46 + 674124e
Closes #30.
Enables the Variational application to operate on a single member of an ensemble cycling setup OR a standalone single-state cycling setup.  Each member has its own job in a separate run directory under `CyclingDA/DATEDIRECTORY/run`.  Adds an EnsembleOfVariational application that carries out the single EDA executable when `useEnsembleOfVariational` is True, the default choice being False.

0d883bd
Fixed a bug in `config/modeldata.csh`, which only would have been triggered if nEnsDAMembers was set larger than 80.

### Testing
Successfully completed 2 full cycles with the following 4 setups
 - OIE120km
   - DAType: 3denvar
   - DAType: eda_3dvar
     - useEnsembleOfVariational: False
       - (1) nEnsDAMembers: 5
       - (2) nEnsDAMembers: 80
     - useEnsembleOfVariational: True
       - (3) nEnsDAMembers: 5
 - O30kmIE120km
   - (4) DAType: 3denvar