Enable multiple EDA instances (#36)
With this PR, the EDA members are configured using `EDASize` and `nDAInstances` instead of `nEnsDAMembers`, which is the product of those two new settings.  When `EDASize == 1` (default), an ensemble of independent Variational applications is conducted just as before.  When `EDASize > 1`, multiple instances of the EnsembleOfVariational application are initiated, each with `EDASize` members.

Users now have the option to configure the minimization algorithm in config/experiment.csh as well.  Although one could use `DRPBlockLanczos` (block EDA) for each of the EDA instances, so far that algorithm is not working for our 120km benchmark experiment.  Full testing/investigation remains for future work.

Closes #33