Use default ncar_pylib to avoid errors (#40)
Fixes bug reported by @ibanos90 in #38

When `ncar_pylib.csh` or `ncar_pylib` (bash) are sourced, they take any command-line arguments that are given to the parent scripts where the source occurs.

For example,


`script0.csh`:
```
./script1.csh arg1
```

`script1.csh`:
```
source ncar_pylib.csh
```

When `script0.csh` is executed,
```
./script0.csh
```
ncar_pylib.csh will receive the argument `arg1`.  The same is true through multiple layers of `source`.  The arguments to `ncar_pylib` have very specific meanings.  Here the solution is to always give `ncar_pylib` a single argument of `default`, which overrides the arguments given to the parent script.

This issue showed up recently, because CISL removed the previously default behavior of propagating the login environment to job scripts.  So the ncar_pylib environment on a login node no longer transfers to the job environment.  We could instead issue the -V PBS flag to `VerifyObs` and `VerifyModel` job scripts, which I have confirmed also works.  However, it is better to declare the entire environment needed for those tasks instead of depending on the user to set up the environment before submitting the cylc suite.