Enable forecast precision to be user-configurable (#41)
A new option, `forecastPrecision`, is added to `config/experiment.csh`, which allows the user to select the precision of MPAS-Atmosphere forecast output.  The Variational and HofX jobs are similar configured to read the same precision.  The Variational job will write analysis fields with the same precision as the input background files.  The default precision is `single`.

As discussed in #22, the problem is mostly solved by setting `precision=single` in all `streams.atmosphere` files.  The one exception is for the first cycle, where the Variational and RTPP jobs MUST output the same precision as is used in the first cycle background file, because the analysis files are initially created as copies of the backgrounds with the same precision.  There are automated precision checks in PrepVariational.csh and RTPPInflation.csh that ensure the correct `analysisPrecision` is used.

Closes #22 

Tests conducted: completed 2 cycles of `eda_3denvar-60-iter_NMEM5_RTPP0.80_LeaveOneOut_OIE120km` for two experiments.  One with `forecastPrecision` set to `single` and another with `forecastPrecision` set to `double`.