Feature/FirstFC_coldstart (#54)
This PR enables MPAS-Workflow to create cold start conditions from an external GFS analysis (cold start), in addition to using a pre-generated first forecast (warm start) and to reinitiate the cycling or the suite from any cycle (restart).

**Changes:**
A new variable "InitializationType" is defined in the config/experiment.csh script which can take values ColdStart/WarmStart/ReStart (note that ReStart is meant for the workflow when restarting from a cycle that had previously failed or stopped for any reason). In this PR is ensured that the initial cycle is the same for both initialization types, cold and warm start. In order to do that, several changes were needed in the config/modeldata.csh script in which the first date is used instead of using the previous one. The next date is also introduced.

In addition, part of the SetupWorkflow.csh script was moved to the "GetWarmStartIC.csh" script, which is the one that creates the link/copy of the pre-generated first forecast. Other changes in SetupWorkflow.csh script were necessary to create the corresponding scripts for each initialization type. The forecast.csh script was also modified to allow its use for the first forecast generation.

Closes https://github.com/NCAR/MPAS-Workflow/issues/51