Use WPS ungrib to decode GFS analysis on RDA (#59)
This PR enables the use of WPS ungrib to decode the GFS analysis in grib format to generate the cold start initial conditions for the MPAS model forecast. GFS analysis on RDA archives on Cheyenne are used instead of pre-downloaded and ungribbed GFS analyses. This allows to run experiments for any date available on RDA archive. Nevertheless, this is still only working for the data period of the pre-generated data since we are using 3denvar and it needs the ensemble forecasts for the covariance matrix. With the implementation of 3dvar in the future, we can then run experiments for any date.

**Changes**:
- The UngribColdStartIC.csh script was created to ungrib the GFS analysis using the ungrib executable. The corresponding namelist was added to the folder config/mpas/init;
- The static data files specification in config/modeldata.csh was modified in order to specify the data source according to the initialization type;
- The if statement for the initialization type in the forecast.csh script was modified in order to specify the correct initial state when using cold start for the first date or any other date. The option to update the sea ice was added in this statement (this may change with the surfaces generation)

**Caveats**:
For now, the sea ice is not updated in this application. It will be implemented in the near future.

Closes https://github.com/NCAR/MPAS-Workflow/issues/53