Add observation preprocessing from RDA (#63)
**Description**:
This PR adds the preprocessing of RDA archived NCEP BUFR observations to IODA-v2 format and their assimilation into the workflow. Jamie's obs2ioda converter (https://github.com/jamiebresch/obs2ioda) is used to convert prepbufr, satwnd, satwind, and 1bamua (default observations in the list provided for now, others can be uncommented and converted as well) observations and then the ioda-upgrade.x executable is used to get the correct string station_id and string variable_names of the preburf observations. This allows to run experiments for any date using the ColdStart initialization type and it can also be used for WarmStart cases. The preprocessing of the observations listed as default takes around 2 minutes of walltime. The walltime increases as more observations are added, but still is less than 10 minutes even with IASI observations.

**Changes**:
- A new script based on rda_obs2ioda.csh (Jamie's script) was added to convert the observations to IODA-v2 and the folder with the new observations is located in the experiment run directory. 
- Adjustments were made in PrepJEDI.csh to account for the proper data source, taking into account if the option to preprocess the observations is on/off.

**Test**:
- Various tests were conducted:
 Cold Start and the new converted observations for different dates ==> OIE120km, 3dvar 
 WarmStart using pre-generated observations and the new converted observations ==> OIE120km, 3dvar, 3denvar
 
**Caveats**:
- The model space verification still is not working for other dates because of the need to generate the GFS analysis for each cycle for the new dates. This will be the subject of future developments.

Closes https://github.com/NCAR/MPAS-Workflow/issues/56