Update initial sea surface fields (#85)
### Description
This PR updates the initial condition with the SST and XICE from the GFS analysis valid at the initial condition time. This is added for the ColdStart initialization type, therefore it can be used for any date and the real time run. The main change is in `forecast.csh` where the initial condition file is updated instead of the forecast output files. In addition, the observation processing dependencies were adjusted so that they can be downloaded and converted to IODA outside the cycling dependency.

### Issue closed
Closes https://github.com/NCAR/MPAS-Workflow/issues/84

### Tests completed
 - [x] 3dvar_OIE120km_WarmStart
 - [x] 3denvar_OIE120km_WarmStart
 - [x] 3dvar_OIE120km_ColdStart
 - [x] 3denvar_O30kmIE60km_WarmStart
 - [x] eda_OIE120km_WarmStart