Reorganize drive.csh using included `.rc` files (#87)
### Description
`drive.csh` is re-organized by utilizing a hierarchy of included `.rc` files in the new `include` directory.  The primary include files are:
- `include/criticalpath.rc`: controls all elements of the critical path for all 4 `CriticalPathType` options
and 2 `InitializationType` options.  Allows for re-use of `include/forecast.rc` and `include/da.rc`
according to the user selections.  Those latter two scripts describe all the intra-forecast and
intra-da dependencies, respectively, independent of tasks in other categories.
- `include/verification.rc`: describes the dependencies between `HofX`, `Verify*`, `Compare*`, and other
kinds of tasks that produce verification statistics files.  It includes dependencies on
`forecast` and `da` tasks that produce the data to be verified.  Multiple aspects of verification
are controlled via the `workflow` section of the scenario configuration. Full descriptions of all
verification options are available in `scenarios/base/workflow.yaml`.
- `include/tasks.rc`: describes all cylc tasks that can be selected under the `[[dependencies]]` node of
`drive.csh`, all of which are described in either `criticalpath.rc`, `verification.rc`, or files
included therein.

`README.md` and `scenarios/base/workflow.rc` are updated to account for the new organization.

### Issue closed

Closes #50 

### Tests completed
 - [x] 3dvar_OIE120km_WarmStart
 - [x] 3denvar_OIE120km_WarmStart
 - [x] 3dvar_OIE120km_ColdStart
 - [x] 3denvar_O30kmIE60km_WarmStart
 - [x] eda_OIE120km_WarmStart