Quality control flags in YAML stubs and VarBC for AMSUA (#92)
### Description
This PR is a follow up of #89 in which the feature branch is renamed and changes are made accordingly following that PR's comments/suggestions.

Here, the variable `biasCorrection` gets only the values `True` or `False` indicating that bias correction is not applied for pre-generated periods (PANDACArchive) since the observations are already bias corrected, and that variational bias correction is applied otherwise. The `biasCorrection` settings is available in both `variational` and `hofx` appications, but is only implemented so far for `variational`.  Setting `biasCorrection` to True for `variational` also turns on variational bias correction coefficient updating.  That new functionality enables the writing of the satellite bias correction coefficients at each cycle and uses the coefficients calculated from the previous cycle as prior estimates.

In addition, the YAML stubs are split and each part is added to a corresponding folder (base, filters, bias, filtersWithBias) for variational and hofx applications. In this way, a great flexibility is allowed in the sense of constructing the full application YAML from a modular set of sub-yamls.

Additionally, `satwnd` observations are enabled for both `variational` and `hofx` applications.  `satwnd` observations will only be used when `observations.resource` is set to either of `GladeRDAOnline` or `NCEPFTPOnline`, such that observations are converted to IODA format from BUFR.

### Issue closed
Closes #43
Closes #82
Closes #86

### Tests completed
 - [x] 3dvar_OIE120km_WarmStart
 - [x] 3denvar_OIE120km_WarmStart
 - [x] 3dvar_OIE120km_ColdStart
 - [x] 3denvar_O30kmIE60km_WarmStart
 - [x] eda_OIE120km_WarmStart