Fix channel selection for NCDIAG AMSUA variational (#96)
### Description
This PR fixes a bug in the channel selection for NCDIAG AMSUA variational. In order to account for the limited channel set for AMSUA instruments used in `PANDACArchive` experiments, the `Useflag check` was added to the base filters. This allows to use the same base yaml files for experiments using `PANDACArchive` or `*Online` data. The channel selection in the check follows the used before the last PR #92.

### Issue closed
Closes #95

### Tests completed
 - [x] 3dvar_OIE120km_WarmStart
 - [x] 3denvar_OIE120km_WarmStart
 - [x] 3dvar_OIE120km_ColdStart
 - [x] 3denvar_O30kmIE60km_WarmStart
 - [x] eda_OIE120km_WarmStart