Add model-space ensemble spread to diagnostics (#99)
### Description

Utilizes new diagnostic functionalities in mpas-jedi for model-space ensemble spread.  Includes modifications to the RTPP task in order to allow the analysis ensemble spread to be diagnosed before the relaxation is applied.

Adds `retain*` options in the `variational`, `hofx`, and `rtpp` applications to allow users to decide whether to retain re-creatable data files and incur the extra disk space requirement.  The default for `variational.retainObsFeedback` and `hofx.retainObsFeedback` is True, then set to False for all EDA scenarios.  The default for `rtpp.retainOriginalAnalyses` is False.

### Issue closed

Closes #94 

### Related PR's

- https://github.com/JCSDA-internal/mpas-jedi/pull/712

### Tests completed
 - [x] 3dvar_OIE120km_WarmStart
 - [x] 3denvar_OIE120km_WarmStart
 - [x] 3dvar_OIE120km_ColdStart
 - [x] 3denvar_O30kmIE60km_WarmStart
 - [x] eda_OIE120km_WarmStart