Update to 04MAY2022 build (#106)
### Description
Use a recent build that includes the new IODA IO pool feature, which allows for the ObsSpace writer to use a subset of the total processors for writing.  Currently each writer process generates a single file.  Previously each process associated with a simulation produced its own file, e.g., 36 files per observation type corresponding to 36 processors in our 120km HofX.  ObsSpace file storage is reduced as follows for files with the `obsout_` prefix produced by HofX in the first cycle of the 2018 experiment:
- Before IO pool feature (36pe): 459MB
- IO pool, default max pool size of 10: 240MB
- IO pool, max pool size of 2: 182MB

Timings for oops::ObsSpace::save, i.e., the writer:
- Before IO pool feature (36pe): 5.4s
- IO pool, default max pool size of 10: 7.9s
- IO pool, max pool size of 2: 25.4s

The GeoVaLs and ObsDiagnostics files still take up a large amount of disk space, but are only produced for a subset of radiance instruments in order to perform cloud-dependent binning.  If we only saved the the ObsSpace files, keeping them around for 80 members over 120 cycles would cost 1.7TB with a max pool size of 2, a 4x reduction over the previous IODA code.  For EDA experiments, it would be reasonable to turn off the cloud-dependent binning that uses GeoVaLs and ObsDiagnostics, then save the ObsSpace files.

### Issue closed

Closes #104 

### Tests completed
 - [x] 3dvar_OIE120km_WarmStart
 - [x] 3denvar_OIE120km_WarmStart
 - [x] 3dvar_OIE120km_ColdStart
 - [x] 3denvar_O30kmIE60km_WarmStart
 - [x] eda_OIE120km_WarmStart