Update `observers` YAML sections (#108)
### Description
Modifications to observations layout in YAML corresponding to source code changes.  The static build is also updated.  Additionally, the number of nearest neighbors is decreased from the default value (4) to 3 in order to capture a small savings in wall time with minimal loss of accuracy.

The wall time for RTPP in the 120km EDA test was too short.  The per-member wall time is increased for both 120km and 60km meshes.

The wall-time and memory for the 60km EDA are set directly in the scenario yaml according to recent benchmarking.

### Issue closed

Closes #107

### Tests completed
 - [x] 3dvar_OIE120km_WarmStart
 - [x] 3denvar_OIE120km_WarmStart
 - [x] 3dvar_OIE120km_ColdStart
 - [x] 3denvar_O30kmIE60km_WarmStart
 - [x] eda_OIE120km_WarmStart