Update web address to download observations and add the download of GDAS analyses (#127)
### Description
This PR updates the web address from where observations for the real-time run are downloaded (see https://github.com/JCSDA-internal/ioda-converters/pull/1009). It also adds a task to download GDAS analyses (atm, sfc, and sfluxgrb) in real-time from the NCEP FTP. New scenarios are also added to run experiments assimilating IASI with a 30kmIE60km dual mesh and 3denvar, ABI with 120km and 3denvar, and AHI with 120km and 3denvar. A bug in the variable `yymmdd` in `GetObs.csh` and `GetGFSanalysis.csh` is fixed. This bug is due to the sourcing of `config/workflow` in `getCycleVars.csh` which re-writes `yymmdd` with the value from the `firstCyclePoint`.

### Tests completed
 - [x] 3dvar_OIE120km_WarmStart
 - [x] 3denvar_OIE120km_WarmStart
 - [x] 3dvar_OIE120km_ColdStart
 - [x] 3denvar_O30kmIE60km_WarmStart
 - [x] eda_OIE120km_WarmStart