Update variational resources, fix bugs in 3dvar_O30kmIE60km_ColdStart (#138)
### Description

There have been many improvements in mpas-jedi recently that have reduced wall-time and memory usage. This PR adjusts memory and wall times for `variational`, with special attention on `30km.60km` settings.

Also fix some bugs in the `3dvar_O30kmIE60km_ColdStart` scenario so that forecasts are successful and it is comparable to other `30km.60km` scenarios.

The job resource settings for `eda_OIE60km_WarmStart` are now identical to the defaults.  So the special settings in the scenario YAML are removed.

### Issue closed

Closes #137

### Tests completed
 - [x] 3dvar_OIE120km_WarmStart
 - [x] 3denvar_OIE120km_WarmStart
 - [x] 3dvar_OIE120km_ColdStart
 - [x] 3dvar_O30kmIE60km_ColdStart
 - [x] 3denvar_O30kmIE60km_WarmStart
 - [x] eda_OIE120km_WarmStart