Add ABIAHI scenario with VarBC, increase walltime for 3dvar 120-km, add amsua_aqua to convertToIODAObservations (#133)
### Description
This PR adds a new scenario to run experiments with ABI and AHI observations, 120-km mesh and 3denvar with variational bias correction. The individual scenarios for ABI and AHI are removed. Here, it is also added an if statement designating the prepbufr files under the prepnr folder on RDA as default and keeps the option for the prep48h files in case the prepnr files are not available. This is particularly important for the RealTime run since the new web link contains only 2 days of data. Amsua_aqua (airsev) is added to the list of observations to be converted to IODA (last 2 days available at https://ftpprd.ncep.noaa.gov/data/nccf/com/obsproc/prod/ and until 20201207 on RDA on Cheyenne) and amsua_metop-c is added to the list of instruments allowing bias correction. Because amsua_metop-c was not yet on that list, those observations were not being assimilated. Finally, the walltime is increased for runs using 3dvar data assimilation with 120-km, especially for the real-time run.

### Tests completed
 - [x] 3dvar_OIE120km_WarmStart
 - [x] 3denvar_OIE120km_WarmStart
 - [x] 3dvar_OIE120km_ColdStart
 - [x] 3denvar_O30kmIE60km_WarmStart
 - [x] eda_OIE120km_WarmStart