Always use valid date for data retrieval (#150)
### Description
Fixes a bug introduced in #145 that would have impacted extended forecast verification when observations and external analyses are retrieved online in the workflow, such as in real-time applications.  The date strings for file retrieval now use the valid date instead of the cycle date.  This would not have impacted the 2018 historical period applications or background forecast verification (0-hr time offset between valid and cycle dates).

### Issue closed

Closes #149 

### Tests completed
Only ColdStart tests and RealTime application are impacted
 - [x] 3dvar_OIE120km_ColdStart
 - [x] 3dvar_O30kmIE60km_ColdStart