Update env-setup scripts for more general usage (#139)
### Description
The current `env-setup` scripts are not sensitive to individual users' `.tcshrc`, `.bashrc`, or other similar login settings files.  This PR introduces measures in `env-setup` to ensure that users load the modules expected by the workflow.  Additional modules are up to the user after that.

### Issue closed

None

### Tests completed
Used for setting up environment for real experiments