Add scenario to download GDAS analyses in real-time and other small modifications (#156)
### Description
This PR adds a separate scenario for the downloading of GDAS analyses (atm, sfc, and sfluxgrb) in near real-time from the NCEP FTP. It adds the necessary task specification and critical path for this stand-alone run that only downloads those data. Also, in this PR are included small changes in the computational requirements of the scenario to run 3denvar OIE120km experiments using 6h IASI observations. The `Retry` option to search for new observations and GFS analyses was also modified to 10minutes instead of 5minutes in order to decrease the frequency this request is submitted to Casper. 

Additionally, it is added the possibility to use the GFS 6-h forecast from previous cycle as the "GFS 0-h forecast" for current cycle, for model space verification purposes. While running a few days the real-time run using RDA data on Cheyenne, it was found out that for some specific days and synoptic hours there is missing the GFS 0-h forecast that we use for the model space verification of MPAS 6-h forecast. Therefore, this will only be of relevance for retrospective experiments using raw data. This was solved by adding an if statement in the script  `GetGFSAnalysisFromRDA.csh` which will allow us to use the previous 6-h forecast.

### Tests completed
 - [x] 3dvar_OIE120km_WarmStart
 - [x] 3denvar_OIE120km_WarmStart
 - [x] 3dvar_OIE120km_ColdStart
 - [x] 3dvar_O30kmIE60km_ColdStart
 - [x] 3denvar_O30kmIE60km_WarmStart
 - [x] eda_OIE120km_WarmStart