Bug fix in gdasfile list and log file name in ObsToIODA (#159)
### Description
This small PR fixes a bug when the ObsToIODA task is executed in the same directory more than once. The loop over the `*gdas*` files is improved and the gdas prefix is removed from the naming of the log files. The deletion of previous log files is also added. I re-run the tests with ColdStart initialization twice and no error was found.

### Issue closed
Closes #148 

### Tests completed
 - [x] 3dvar_OIE120km_WarmStart
 - [x] 3denvar_OIE120km_WarmStart
 - [x] 3dvar_OIE120km_ColdStart
 - [x] 3dvar_O30kmIE60km_ColdStart
 - [x] 3denvar_O30kmIE60km_WarmStart
 - [x] eda_OIE120km_WarmStart