BUMP yaml updates (#180)
### Description
Companion to https://github.com/JCSDA-internal/mpas-jedi/pull/811.  YAML formats are changed for climatological B and localization.  Refactoring only.

Build is updated to 06JAN2023, which will be updated on the merge date to be consistent with develop code on that date.

Some additional changes are made in `ExternalAnalysisToMPAS.csh` to improve its restart behavior when failing on the first attempt.

### Issue closed

Closes #174 

### Tests completed
 - [x] 3dvar_OIE120km_WarmStart
 - [x] 3denvar_OIE120km_WarmStart
 - [x] 3dvar_OIE120km_ColdStart
 - [x] 3dvar_O30kmIE60km_ColdStart
 - [x] 3denvar_O30kmIE60km_WarmStart
 - [x] eda_OIE120km_WarmStart