Enable flexibility for non-bundle Forecast build (#217)
### Description
A scenario YAML option named `build.forecast directory` is added that allows users to simply set the directory where the `init_atmosphere_model` and `atmosphere_model` executables are located.  The default is `bundle`, which uses the `mpas bundle` directory and JEDI environment.

The environment for the non-bundle executable is defined in `config/environmentForecast.csh`, including the mpi command (e.g., mpiexec_mpt) that should be used.

### Issue closed

Closes #216 

### Tests completed
#### Tier 1:
 - [x] 3dvar_OIE120km_WarmStart
 - [x] ForecastFromGFSAnalysesMPT
 
As long as a bundle forecast and a non-bundle forecast executable are successful, then this PR is correct.  No need to run other Tier 1 tests.