Add share queue as option for single processor jobs (#221)
### Description
Casper compute nodes are down today, which means we cannot use the high-throughput queue.  I tried replacing that queue selection with the share queue, and it worked.  This gives everybody the option to run whenever Casper is under maintenance.

### Issue closed

None

### Tests completed
Successfully ran an experiment with `hpc.SingleProcQueue` set to `share`.