Fix CleanHofX task failure (#293)
### Description
This PR fixes the failure in `CleanHofXBG-0hr00` task when a CYLC suite is reload/replayed after it stopped. The task fails because `ExperimentDirectory` is not defined. This is fixed by sourcing `config/auto/experiment.csh` (which contains the `ExperimentDirectory` variable definition) in the beginning of the bin/CleanHofX.csh. Thanks May Wong for this fix!  This was also fixed in `bin/comparemodel.csh` which had the same issue.

### Issue closed
Closes https://github.com/NCAR/MPAS-Workflow/issues/290

### List of modified files:
M bin/CleanHofX.csh
M bin/comparemodel.csh

### Tests completed
- Ran test with suite that had failed and now is successful