Fix linking satbias file if no obs file exists (#294)
### Description
This PR reviews linking satbias files when the observation file does not exist. An HDF5 error arises because it cannot open the file, but that's actually because the file does not exist. In the `PrepJEDI.csh`, we link the satbias file from previous cycle and use the variable `${CyclingDADir}` to indicate where to link it. However, `CyclingDADir` already includes the date information (as defined in `getCycleVars.csh`), then `${CyclingDADir}/${thisValidDate}/dbOut` produces a wrong path. The cleaner solution is to use the variable `DAWorkDir` that does not include date information and keep `${thisValidDate}` to ensure the targeted path is correct. This is equivalent to `${CyclingDADir}/dbOut`. 

### Issue closed
Closes https://github.com/NCAR/MPAS-Workflow/issues/292

### Tests completed
- Tested in suite that failed because satbias files was not found