Activate Cloud_Seeding parameter and remove 'max frame size' from obsdatain (#301)
### Description
This PR follows up on recent changes in UFO to avoid `Tb_clear` getting zero values. This is done by activating the `Cloud_Seeding` parameter in the YAML section for the CRTM cloudy forward operator. Visual results of this functionality are provided in https://github.com/JCSDA-internal/ufo/pull/3264.

In addition, the `max frame size` is removed from `obsdatain` in several ObsSpaces. This option was removed from the YAML configuration in https://github.com/JCSDA-internal/ioda/pull/1079. These changes come with an updated build code of the develop branches.

### Issue closed
None

### Tests completed
#### Tier 1:
 - [x] 3dvar_OIE120km_WarmStart
 - [x] 3denvar_OIE120km_IAU_WarmStart
 - [x] 3dvar_OIE120km_ColdStart
 - [x] 3dvar_O30kmIE60km_ColdStart
 - [x] 3denvar_O30kmIE60km_WarmStart
 - [x] eda_OIE120km_WarmStart
 - [x] getkf_OIE120km_WarmStart
 - [x] ForecastFromGFSAnalysesMPT